indexer: Remove old compacted file only if it exists
When compaction is triggered for an already compacted file,
forestdb does a file move to new filename. Hence, old compacted
file may go missing.

+++ secondary/indexer/forestdb_slice_writer.go
@@ -623,7 +623,10 @@ func (fdb *fdbSlice) Compact() error {
 		return err
 	}
 
-	err = os.Remove(fdb.currfile)
+	if _, e := os.Stat(fdb.currfile); e == nil {
+		err = os.Remove(fdb.currfile)
+	}
+
 	fdb.currfile = newpath
 	return err
 }
